[REF] move CLAChecker class in a separate module
and initialize db in main()
